add sumPoint

+++ src/main/java/com/zjwm/wyx/login/service/UserService.java
@@ -2,9 +2,6 @@ package com.zjwm.wyx.login.service;
 
 import com.zjwm.wyx.login.entity.UserEntity;
 
-import java.util.List;
-import java.util.Map;
-
 /**
  * 用户
  * 
@@ -14,7 +11,7 @@ import java.util.Map;
  */
 public interface UserService {
 	
-	UserEntity queryObject(Long userId);
+	UserEntity queryObject(int userId);
 
 	UserEntity queryByMobile(String mobile);
 
@@ -24,7 +21,7 @@ public interface UserService {
 	 * @param password  密码
 	 * @return          返回用户ID
 	 */
-	long login(String mobile, String password);
+	int login(String mobile, String password);
 
 	void save(String mobile, String password);
 }
